Introduce support to WASM compilation by avoiding std lib
This commit introduce compilation features to allow this library to be used on
`wasm32-unknown-unknown` target.

The `crossbeam_epoch` library already have support to compile on a `no_std`
environment, as long as we don't use the global `crossbeam_epoch::epoch::pin`
lock, which is feature-gated for `std` environments.

In order to compile this library for WASM, the following changes were performed:
- Instead of a global epoch value, we now use a per `Recloser` epoch lock.
- We feature-gate the AsyncRecloser behind a `feature=async` feature flag.

To avoid breaking users, the `async` flag is provided as a `default` feature.
Now, this library can be compiled to a WASM environment using:

```sh
cargo build --release --target wasm32-unknown-unknown --no-default-features
```

And now can be included like this on a WASM library:

```toml
[dependencies]
recloser = { version = "*", default-features = false }
```